+++ bridge-sdk/bridge-clients/near-bridge-client/src/btc_connector.rs
@@ -115,7 +115,7 @@ impl NearBridgeClient {
     pub async fn sign_btc_transaction(
         &self,
         btc_pending_id: String,
-        sign_index: usize,
+        sign_index: u64,
         transaction_options: TransactionOptions,
     ) -> Result<CryptoHash> {
         let endpoint = self.endpoint()?;
